test personaNotFound en modificar_telefono

+++ src/main/java/com/babel/vehiclerentingapproval/services/impl/PersonaServiceImpl.java
@@ -211,19 +211,21 @@ public class PersonaServiceImpl implements PersonaService {
      * @return void
      */
     @Transactional
-    public void modificarTelefono(Persona persona) {
-        List<TelefonoContacto> telefonos = persona.getTelefonos();
-        List<TelefonoContacto> telefonosAntiguos = telefonoMapper.listarTelefonos(persona.getPersonaId());
+    public void modificarTelefono(Persona persona) throws PersonaNotFoundException{
 
-        //Borramos telefonos pertenecientes al usuario
-        for (int i = 0; i < telefonosAntiguos.size(); i++) {
-            this.telefonoMapper.deleteTelefono(persona.getPersonaId(), telefonosAntiguos.get(i));
-        }
-        //Añadimos los telefonos del usuario
-        for (int i = 0;i<telefonos.size();i++) {
-            this.telefonoMapper.addTelefono(telefonos.get(i),persona);
-        }
+        if(existePersona(persona.getPersonaId())) {
+            List<TelefonoContacto> telefonos = persona.getTelefonos();
+            List<TelefonoContacto> telefonosAntiguos = telefonoMapper.listarTelefonos(persona.getPersonaId());
 
+            for (int i = 0; i < telefonosAntiguos.size(); i++) {
+                this.telefonoMapper.deleteTelefono(persona.getPersonaId(), telefonosAntiguos.get(i));
+            }
+            for (int i = 0; i < telefonos.size(); i++) {
+                this.telefonoMapper.addTelefono(telefonos.get(i), persona);
+            }
+        }else{
+            throw new PersonaNotFoundException();
+        }
     }
 
     /**
